🐛(defaults) fix settings circular import
The defaults pattern used in this repository implies loading django
settings, but, we also need to import the
filter_configurable_lti_consumer in django settings. As xblocks force us
to import XBlock from the module's root, we had a circular importation
issue with django settings.

We now set default configuration at runtime if it has not been defined
in the project's settings.