🐛(circleci) fix pypi distribution cache key
The pypi packages were saved to cache using the checksum of the setup.cfg
file as a key. This lead circleci to push the wrong package on our release
1.0.0-rc.2 because the PR to bump the version number was prepared before
merging everything that we wanted to add to this PR. The release commit
was then rebased on master but it was too late, the cache in circleCI
was already set with an incomplete tar.gz.

Using the commit hash as cache key for distribution packages ensures that
it will vary each time the code varies.