✨(configuration) allow defining dynamic launch urls
In some cases the LTI service we are calling requires a unique launch url
for each XBlock calling the service. Indeed, the LTI specification requires
that the XBlock be identified by its url, not by its resource_link_id.

We use exrex to generate a random launch url respecting a regex. This can
for example be used to generate launch urls of the form:
https://marsha.education/lti-video/d8edd6fb-09e2-405d-e2e2-6f41734eac85
by defining the following regex:
https://marsha.education/lti-video/[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}