🐛(configuration) allow setting visible fields effectively
The value of fields was taken from our settings (and not from Mongodb
which is the default behavior). When a field is visible (not hidden)
the user expects to see the effects of changes he makes. For this,
we should take field values from our settings only for hidden fields,
and stick to the normal behavior of reading field values from Mongodb
for visible fields.